ai-engine/requirements.txt
fastapi
uvicorn
groq
python-dotenv
numpy